requirements.txt
# To install every package `pip install -r requirements.txt`
pillow
pygame

